feat: A-Enums
