+++ src/main/java/Tony.java
@@ -1,11 +1,11 @@
-import java.util.ArrayList;
-import java.util.List;
+import tasks.Task;
+import tasks.TaskType;
 import java.util.Scanner;
 import java.util.Arrays;
 
 public class Tony {
     public static void main(String[] args) {
-        ToDoList lst = new ToDoList();
+        TodoList lst = new TodoList();
         Scanner scanner = new Scanner(System.in);  // Create a Scanner object
         greeting();
         String input = scanner.nextLine();
@@ -23,17 +23,17 @@ public class Tony {
                     lst.unmark(secondWord);
                 } else if (firstWord.equals("todo")) {
                     String result = String.join(" ", Arrays.copyOfRange(words, 1, words.length));
-                    Task toDo = new Todo(result);
+                    Task toDo = new TaskFactory().createTask(TaskType.TODO, result);
                     lst.add(toDo);
                 } else if (firstWord.equals("deadline")) {
                     String result = String.join(" ", Arrays.copyOfRange(words, 1, words.length));
                     String[] parts = result.split("/");
-                    Task deadline = new Deadline(parts);
+                    Task deadline = new TaskFactory().createTask(TaskType.DEADLINE, parts);
                     lst.add(deadline);
                 } else if (firstWord.equals("event")) {
                     String result = String.join(" ", Arrays.copyOfRange(words, 1, words.length));
                     String[] parts = result.split("/");
-                    Task event = new Event(parts);
+                    Task event = new TaskFactory().createTask(TaskType.EVENT, parts);
                     lst.add(event);
                 } else if (firstWord.equals("delete")) {
                     String secondWord = words[1];
@@ -69,144 +69,5 @@ public class Tony {
         System.out.println("_______________________\n");
     }
 
-    static class ToDoList {
-        List<Task> list = new ArrayList<>();
-
-        public void add(Task item) {
-            list.add(item);
-            int numberOfTasks = list.size();
-            line();
-            System.out.println("Got it dawg. I've added this task: \n");
-            System.out.println(item.toString() + "\n");
-            System.out.println("Now you got "+ numberOfTasks + " tasks fam \n");
-            line();
-        }
-
-        public void mark(String input) {
-            try {
-                int index = Integer.parseInt(input);
-                list.get(index - 1).markAsDone();
-                line();
-                System.out.println("Marked item " + index + " as done dawg.");
-                line();
-            } catch (NumberFormatException | IndexOutOfBoundsException e) {
-                line();
-                System.out.println("Invalid input for 'mark' command dawg.");
-                line();
-            }
-        }
-
-        public void unmark(String input) {
-            try {
-                int index = Integer.parseInt(input);
-                list.get(index - 1).markAsUndone();
-                line();
-                System.out.println("Unmarked item " + index + " as done.");
-                line();
-            } catch (NumberFormatException | IndexOutOfBoundsException e) {
-                line();
-                System.out.println("Invalid input for 'unmark' command.");
-                line();
-            }
-        }
-
-        public void delete(String input) {
-            try {
-                int index = Integer.parseInt(input);
-                String task = list.get(index - 1).toString();
-                list.remove(index - 1);
-                line();
-                System.out.println("Deleted item: \n" + task + "\n");
-                System.out.println("Now you have " + list.size() + " tasks left in the list.");
-                line();
-            } catch (NumberFormatException | IndexOutOfBoundsException e) {
-                line();
-                System.out.println("Invalid input for 'delete' command.");
-                line();
-            }
-        }
-
-        public void print() {
-            line();
-            System.out.println("Here are the tasks in your list: \n");
-            for (int i = 0; i < list.size(); i++) {
-                System.out.println(list.get(i).toString() + "\n");
-            }
-            line();
-        }
-    }
 }
 
-class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void markAsUndone() {
-        isDone = false;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + (isDone ? "X" : " ") + "] " + description;
-    }
-}
-
-class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-        if (description.equals("")) {
-            throw new IllegalArgumentException("Should have more description dawg");
-        }
-
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
-
-class Deadline extends Task {
-    private String dueDate;
-
-    public Deadline(String[] parts) {
-        super(parts[0]);
-        if (parts.length < 2) {
-            throw new IllegalArgumentException("Invalid input for deadline task. Expected: <description> /by <dueDate>");
-        }
-        this.dueDate = parts[1];
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (" + dueDate + ")";
-    }
-}
-
-class Event extends Task {
-    private String from;
-    private String to;
-
-    public Event(String[] parts) {
-        super(parts[0]);
-        if (parts.length < 3) {
-            throw new IllegalArgumentException("Invalid input for event task. Expected: <description> /at <from> <to>");
-        }
-        this.from = parts[1];
-        this.to = parts[2];
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (" + from + " " + to + ")";
-    }
-}
