+++ src/main/java/tasks/Todo.java
@@ -1,0 +1,22 @@
+package tasks;
+
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+        this.type = TaskType.TODO;
+        if (description.equals("")) {
+            throw new IllegalArgumentException("Should have more description dawg");
+        }
+
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String getType() {
+        return type.toString();
+    }
+}
